Moved notifications function to separate file
