+++ src/App.tsx
@@ -52,7 +52,7 @@ import { Menu } from "./modals/Menu";
 import { isNewVersionAvailable } from "./data/AppVersion";
 import { configuration } from "./data/AppConfiguration";
 
-import { LocalNotifications } from "@capacitor/local-notifications";
+import { requestPermission, scheduleNotification } from "./utils/notifications";
 
 setupIonicReact();
 
@@ -93,35 +93,6 @@ const App = (props: AppProps) => {
     },
     [i18n],
   );
-
-  const scheduleNotification = () => {
-    LocalNotifications.removeAllDeliveredNotifications()
-      .then(() => {
-        console.log("Old notifications removed");
-
-        const notificationId = Date.now();
-
-        return LocalNotifications.schedule({
-          notifications: [
-            {
-              id: notificationId,
-              title: "Notification!",
-              body: "Time for something important",
-              schedule: { at: new Date(Date.now() + 5000) },
-              sound: "default",
-              smallIcon: "ic_launcher",
-              largeIcon: "ic_launcher",
-            },
-          ],
-        });
-      })
-      .then(() => {
-        console.log("Notification scheduled");
-      })
-      .catch((error) => {
-        console.error("Error handling notifications:", error);
-      });
-  };
 
   function updateCycles(newCycles: Cycle[]) {
     const slicedCycles = newCycles.slice(0, maxOfCycles);
@@ -210,20 +181,6 @@ const App = (props: AppProps) => {
     if (!configuration.features.notifications || !notificationsStatus) {
       return;
     }
-
-    const requestPermission = () => {
-      LocalNotifications.requestPermissions()
-        .then(({ display }) => {
-          if (display !== "granted") {
-            console.log("Notification permission not received");
-            return;
-          }
-          console.log("Notification permission granted");
-        })
-        .catch((err) => {
-          console.error("Error requesting notification permission:", err);
-        });
-    };
 
     requestPermission();
   }, [notificationsStatus]);
